// src/types/reducer-types.ts
import {CartItem, ShippingInfo, User} from "./types"

export interface UserReducerInitialState {
    user: User | null;
    loading: boolean
}

export interface CartReducerInitalState {
    loading: boolean
    cartItems: CartItem[]
    subtotal: number
    tax: number
    shippingCharges: number
    discount: number
    total: number
    shippingInfo: ShippingInfo
}